Moves components to their own directory
