+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { HashRouter } from 'react-router-dom'
-import './index.css'
 import App from './App'
+import './index.css'
 
 ReactDOM.render(
   <React.StrictMode>
